Use square function from the libraries.
An interesting read might be this : https://wiki.haskell.org/Power_function
The underlying problem is that Haskell will not auto-coerce, so it needs
a power-function with a signature Int -> Int -> Int as well as
Double -> Int -> Double as well as ...
We can go a little deeper into it next session if you feel like. There
might be an interesting link to the numeric prelude where they would like
to fix this with multi parameter type classes. This might go a bit too
far right now.